refactoringTask - добавлено разделение на подпакеты для лучшего стуктурирования задач по спринтам
