+++ src/main/java/school/faang/moduleone/sprintone/task_43502/Warrior.java
@@ -1,4 +1,4 @@
-package school.faang.moduleone.task_43502;
+package school.faang.moduleone.sprintone.task_43502;
 
 public class Warrior extends Character {
     private static final int WARRIORS_POWER = 10;
