+++ src/main/java/school/faang/moduleone/sprintone/task_43503/Main.java
@@ -1,4 +1,4 @@
-package school.faang.moduleone.task_43503;
+package school.faang.moduleone.sprintone.task_43503;
 
 import java.util.List;
