+++ src/main/java/school/faang/moduleone/sprintone/task_43504/User.java
@@ -1,4 +1,4 @@
-package school.faang.moduleone.task_43504;
+package school.faang.moduleone.sprintone.task_43504;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
