+++ src/main/java/school/faang/moduleone/sprintone/task_43513/SpellEvent.java
@@ -1,4 +1,4 @@
-package school.faang.moduleone.task_43513;
+package school.faang.moduleone.sprintone.task_43513;
 
 public record SpellEvent(int id, String eventType, String action) {
 }
